Move copy functions to separate file
